+++ retrofit/src/test/java/retrofit/RestAdapterTest.java
@@ -209,8 +209,7 @@ public class RestAdapterTest {
 
   @Test public void asynchronousUsesExecutors() throws Exception {
     Response response = new Response(200, "OK", NO_HEADERS, new TypedString("{}"));
-    when(mockClient.execute(any(Request.class))) //
-        .thenReturn(response);
+    when(mockClient.execute(any(Request.class))).thenReturn(response);
     Callback<Object> callback = mock(Callback.class);
 
     example.something(callback);